sync final fix - test version working

+++ repositoriesLocalSHP/syncSystem.py
@@ -168,9 +168,10 @@ class WateringSync():
         self.layer = change.layer_id
         feature_to_delete = self.get_feature_by_id(change.feature_id)
         
-        self.layer.startEditing()
-        self.layer.deleteFeatures(feature_to_delete.id())
-        self.layer.commitChanges()
+        if feature_to_delete:
+            self.layer.startEditing()
+            self.layer.deleteFeatures(feature_to_delete.id())
+            self.layer.commitChanges()
 
     def save_offline_changes_to_project(self):
        serialized_changes = json.dumps([change.__dict__ for change in self.offline_change_queue])
